[automessage] Xavier'fortune ;)
In Reading [England] there is this thing called the IDR, short for "Inner
Distribution Road", which is bureaucratese for "Big thing that cost a lot
of money and relieves traffic problems, provided all your traffic wants to
orbit the town centre permanently". It's a 2-3 lane dual carriageway that
goes round the town centre. It has lots of roundabouts, an overhead
section, a couple of spare motorway-like exits (that's British motorways --
y'know, the roundabout with the main road going under it), and a thing
called the Watlington Street Gyratory, where you have to get in lane for
your intended destination about three years and two corners before you get
there *with no signposting*. I used to cycle along it every day to get to
school, before I fell off at 35 mph. [Kids! Don't try this at home!]  I
know it well. I believe it is impossible to leave Reading heading west.
        -- (Terry Pratchett, alt.fan.pratchett)